+++ packages/orm/CHANGELOG.md
@@ -1,5 +1,18 @@
 # orchid-orm
 
+## 1.5.0
+
+### Minor Changes
+
+- Change index options: column or expression is required, operator renamed to opclass
+
+### Patch Changes
+
+- f1cd5db: Handle multiple indexes and foreignKeys of the column
+- Updated dependencies
+- Updated dependencies [f1cd5db]
+  - pqb@0.9.0
+
 ## 1.4.22
 
 ### Patch Changes